Refactor: Moved calc function from shoppingList to data and other small refactor stuff
